+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `Down::ChunkedIO#closed?` (@janko-m)
+
 * Add `Down::ChunkedIO#rewindable?` (@janko-m)
 
 * In `Down::ChunkedIO` only create the Tempfile if it's going to be used (@janko-m)